added custom hook and a layer for handling api calls
